Updates issue with Express Checkout error
If the API credentials entered into the admin panel
where incorrect the PayPal API was returning an
expected error, but it was stranding the user on
a useless page.  This has been configured to return
the user to the cart page and display a message to
try another payment method, as well as send an email
to the site admin notifying them about the problem.